made web3-connection.ts export the functions since it kinda made sense
